+++ libft/ft_itoa_base.c
@@ -6,13 +6,13 @@
 /*   By: midrissi <midrissi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/25 17:18:31 by midrissi          #+#    #+#             */
-/*   Updated: 2019/02/05 16:48:53 by midrissi         ###   ########.fr       */
+/*   Updated: 2019/02/06 15:45:12 by midrissi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_itoa_base(long long n, int base, int uppercase)
+char	*ft_itoa_base(intmax_t n, int base, int uppercase)
 {
 	char *nb;
